requirements.txt
python-docx>=0.8.11
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
openai>=1.0.0
python-dotenv>=0.19.0
chromadb>=0.4.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2
spacy>=3.7.0
nltk>=3.8.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
redis>=4.0.0
pymongo>=4.1.1
numpy>=1.24.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
